requirements.txt
yt-dlp
google-genai
python-dotenv
dnspython
#ffmeg?